implementation/exercise-2/app/web2/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
SQLAlchemy==2.0.30
Werkzeug==3.0.3
python-dotenv==1.0.1
